enables execution of quickstart tests in example folder
